Move import logic to ProfileImportService
